updated user authentification
