+++ package.json
@@ -19,12 +19,16 @@
   "homepage": "https://github.com/suryajasper/Food-Bank-Smart-Routes#readme",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "crypto": "^1.0.1",
     "express": "^4.17.1",
     "firebase-admin": "^8.12.1",
     "google-spreadsheet": "^2.0.6",
     "googleapis": "^39.2.0",
     "http": "0.0.1-security",
     "iwanthue": "^1.4.2",
+    "mongodb": "^4.2.2",
+    "mongoose": "^6.1.4",
+    "mongoose-unique-validator": "^3.0.0",
     "node_or_tools": "^1.0.7",
     "socket.io": "^2.3.0",
     "unirest": "^0.6.0",
